moved count cards to ui-elements
